# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3torConnect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type LoginRequestRole =
  (typeof LoginRequestRole)[keyof typeof LoginRequestRole];

export const LoginRequestRole = {
  student: "student",
  tutor: "tutor",
  admin: "admin",
} as const;

export interface LoginRequest {
  email: string;
  password: string;
  role: LoginRequestRole;
}

export interface RegisterStudentRequest {
  name: string;
  email: string;
  password: string;
  phone?: string;
  university?: string;
  admissionType?: string;
  jambScore?: number;
}

export interface RegisterTutorRequest {
  name: string;
  email: string;
  password: string;
  phone?: string;
  university?: string;
  faculty?: string;
  department?: string;
  level?: string;
  aboutYou?: string;
  subjects?: string[];
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  student: "student",
  tutor: "tutor",
  admin: "admin",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  role: UserProfileRole;
  phone?: string | null;
  status: string;
  createdAt: string;
  lastLogin?: string | null;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export type UpdateUserRequestStatus =
  (typeof UpdateUserRequestStatus)[keyof typeof UpdateUserRequestStatus];

export const UpdateUserRequestStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
  rejected: "rejected",
} as const;

export interface UpdateUserRequest {
  status?: UpdateUserRequestStatus;
  name?: string;
  phone?: string;
}

export interface UsersListResponse {
  users: UserProfile[];
  total: number;
  page: number;
  limit: number;
}

export interface TutorProfile {
  id: number;
  userId: number;
  name: string;
  email: string;
  phone?: string | null;
  status: string;
  university?: string | null;
  faculty?: string | null;
  department?: string | null;
  level?: string | null;
  aboutYou?: string | null;
  subjects?: string[] | null;
  hourlyRate?: string | null;
  rating?: string | null;
  totalSessions?: number | null;
  isVerified?: boolean | null;
  applicationDate: string;
}

export interface UpdateTutorRequest {
  university?: string;
  faculty?: string;
  department?: string;
  level?: string;
  aboutYou?: string;
  subjects?: string[];
  hourlyRate?: string;
  isVerified?: boolean;
}

export interface TutorsListResponse {
  tutors: TutorProfile[];
  total: number;
  page: number;
  limit: number;
}

export interface AvailabilitySlot {
  id: number;
  tutorId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  isAvailable?: boolean;
}

export interface AvailabilityListResponse {
  slots: AvailabilitySlot[];
}

export type SetAvailabilityRequestSlotsItem = {
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  isAvailable?: boolean;
};

export interface SetAvailabilityRequest {
  slots: SetAvailabilityRequestSlotsItem[];
}

export type SessionDetailStatus =
  (typeof SessionDetailStatus)[keyof typeof SessionDetailStatus];

export const SessionDetailStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface SessionDetail {
  id: number;
  tutorId: number;
  studentId: number;
  tutorName: string;
  studentName: string;
  subject: string;
  scheduledAt: string;
  durationMinutes: number;
  status: SessionDetailStatus;
  amount?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateSessionRequest {
  tutorId: number;
  studentId: number;
  subject: string;
  scheduledAt: string;
  durationMinutes?: number;
  amount?: string;
  notes?: string;
}

export type UpdateSessionRequestStatus =
  (typeof UpdateSessionRequestStatus)[keyof typeof UpdateSessionRequestStatus];

export const UpdateSessionRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateSessionRequest {
  status?: UpdateSessionRequestStatus;
  notes?: string;
}

export interface SessionsListResponse {
  sessions: SessionDetail[];
  total: number;
  page: number;
  limit: number;
}

export type TransactionItemType =
  (typeof TransactionItemType)[keyof typeof TransactionItemType];

export const TransactionItemType = {
  payment: "payment",
  withdrawal: "withdrawal",
  refund: "refund",
  bonus: "bonus",
} as const;

export type TransactionItemStatus =
  (typeof TransactionItemStatus)[keyof typeof TransactionItemStatus];

export const TransactionItemStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface TransactionItem {
  id: number;
  userId: number;
  sessionId?: number | null;
  type: TransactionItemType;
  amount: string;
  description?: string | null;
  status: TransactionItemStatus;
  createdAt: string;
  userName?: string | null;
}

export interface TransactionsListResponse {
  transactions: TransactionItem[];
  total: number;
  page: number;
  limit: number;
}

export interface MessageItem {
  id: number;
  senderId: number;
  receiverId: number;
  content: string;
  isRead: boolean;
  createdAt: string;
  senderName?: string | null;
}

export interface MessagesListResponse {
  messages: MessageItem[];
}

export interface SendMessageRequest {
  receiverId: number;
  content: string;
}

export interface ConversationItem {
  userId: number;
  userName: string;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export interface ConversationsListResponse {
  conversations: ConversationItem[];
}

export interface AdminStats {
  totalUsers: number;
  totalStudents: number;
  totalTutors: number;
  totalSessions: number;
  totalRevenue: string;
  pendingTutors: number;
  activeSessions: number;
  recentUsers: UserProfile[];
  recentSessions: SessionDetail[];
}

export type ListUsersParams = {
  role?: ListUsersRole;
  status?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  student: "student",
  tutor: "tutor",
  admin: "admin",
} as const;

export type ListTutorsParams = {
  subject?: string;
  status?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListSessionsParams = {
  tutorId?: number;
  studentId?: number;
  status?: string;
  page?: number;
  limit?: number;
};

export type ListTransactionsParams = {
  userId?: number;
  type?: string;
  status?: string;
  page?: number;
  limit?: number;
};

export type ListMessagesParams = {
  otherUserId: number;
};