# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 3torConnect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/student:
    post:
      operationId: registerStudent
      tags: [auth]
      summary: Register a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterStudentRequest"
      responses:
        "201":
          description: Student registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/tutor:
    post:
      operationId: registerTutor
      tags: [auth]
      summary: Register a tutor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTutorRequest"
      responses:
        "201":
          description: Tutor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tutors:
    get:
      operationId: listTutors
      tags: [tutors]
      summary: List tutors
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tutors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorsListResponse"
  /tutors/{id}:
    get:
      operationId: getTutor
      tags: [tutors]
      summary: Get tutor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateTutor
      tags: [tutors]
      summary: Update tutor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTutorRequest"
      responses:
        "200":
          description: Updated tutor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorProfile"
  /tutors/{id}/availability:
    get:
      operationId: getTutorAvailability
      tags: [availability]
      summary: Get tutor availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityListResponse"

    put:
      operationId: setTutorAvailability
      tags: [availability]
      summary: Set tutor availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAvailabilityRequest"
      responses:
        "200":
          description: Updated availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityListResponse"
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a conversation
      parameters:
        - name: otherUserId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesListResponse"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageItem"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [student, tutor, admin]
      required:
        - email
        - password
        - role
    RegisterStudentRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
        university:
          type: string
        admissionType:
          type: string
        jambScore:
          type: integer
      required:
        - name
        - email
        - password
    RegisterTutorRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
        university:
          type: string
        faculty:
          type: string
        department:
          type: string
        level:
          type: string
        aboutYou:
          type: string
        subjects:
          type: array
          items:
            type: string
      required:
        - name
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
      required:
        - user
        - token
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, tutor, admin]
        phone:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - email
        - role
        - status
        - createdAt
...
    TutorProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: string
        university:
          type: string
          nullable: true
        faculty:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        level:
          type: string
          nullable: true
        aboutYou:
          type: string
          nullable: true
        subjects:
          type: array
          items:
            type: string
          nullable: true
        hourlyRate:
          type: string
          nullable: true
        rating:
          type: string
          nullable: true
        totalSessions:
          type: integer
          nullable: true
        isVerified:
          type: boolean
          nullable: true
        applicationDate:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - email
        - status
        - applicationDate
    UpdateTutorRequest:
      type: object
      properties:
        university:
          type: string
        faculty:
          type: string
        department:
          type: string
        level:
          type: string
        aboutYou:
          type: string
        subjects:
          type: array
          items:
            type: string
        hourlyRate:
          type: string
        isVerified:
          type: boolean
    TutorsListResponse:
      type: object
      properties:
        tutors:
          type: array
          items:
            $ref: "#/components/schemas/TutorProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - tutors
        - total
        - page
        - limit
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: integer
        tutorId:
          type: integer
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        isAvailable:
          type: boolean
      required:
        - id
        - tutorId
        - dayOfWeek
        - startTime
        - endTime
    AvailabilityListResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilitySlot"
      required:
        - slots
    SetAvailabilityRequest:
      type: object
      properties:
        slots:
          type: array
          items:
            type: object
            properties:
              dayOfWeek:
                type: integer
              startTime:
                type: string
              endTime:
                type: string
              isAvailable:
                type: boolean
            required:
              - dayOfWeek
              - startTime
              - endTime
      required:
        - slots
...
    MessageItem:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        senderName:
          type: string
          nullable: true
      required:
        - id
        - senderId
        - receiverId
        - content
        - isRead
        - createdAt
    MessagesListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageItem"
      required:
        - messages
    SendMessageRequest:
      type: object
      properties:
        receiverId:
          type: integer
        content:
          type: string
      required:
        - receiverId
        - content
...