# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3torConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilitySlot {
  id: number;
  tutorId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  isAvailable?: boolean;
}