# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3torConnect API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginRequestRole } from "./loginRequestRole";

export interface LoginRequest {
  email: string;
  password: string;
  role: LoginRequestRole;
}