# lib/api-zod/src/generated/types/updateTutorRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3torConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTutorRequest {
  university?: string;
  faculty?: string;
  department?: string;
  level?: string;
  aboutYou?: string;
  subjects?: string[];
  hourlyRate?: string;
  isVerified?: boolean;
}